fixed bug and generated map of holdings by date from holdings object

+++ app/src/main/java/edu/northeastern/group_project_group_8/Holding.java
@@ -5,11 +5,11 @@ import java.time.LocalDate;
 public class Holding {
     String account;
     String asset;
-    int count;
+    long count;
     LocalDate startDate;
     LocalDate endDate;
 
-    public Holding(String account, String asset, int count, LocalDate startDate, LocalDate endDate) {
+    public Holding(String account, String asset, long count, LocalDate startDate, LocalDate endDate) {
         this.account = account;
         this.asset = asset;
         this.count = count;
